+++ src/pages/AuthLayout.tsx
